refactored whole code. I created component folder
