+++ server/readme.md
@@ -1,0 +1,2 @@
+# run
+node server.js
