+++ vite_proj/index.html
@@ -175,7 +175,7 @@
 
     </footer>
 
-    <script type="module" src="/src/app.ts"></script>
+    <script type="module" src="/src/main.ts"></script>
 
 </body>
