+++ vite_proj/readme.md
@@ -1,0 +1,2 @@
+# run
+npm run dev
