+++ vite_proj/src/components/Counter.ts
@@ -1,0 +1,9 @@
+export function setupCounter(element: HTMLButtonElement) {
+  let counter = 0;
+  const setCounter = (count: number) => {
+    counter = count;
+    element.innerHTML = `count is ${counter}`;
+  };
+  element.addEventListener('click', () => setCounter(counter + 1));
+  setCounter(0);
+}
