+++ vite_proj/src/main.ts
@@ -1,24 +1,15 @@
-import './style.css'
-import typescriptLogo from './typescript.svg'
-import viteLogo from '/vite.svg'
-import { setupCounter } from './counter.ts'
+import './styles/style.css';
+import { fetch_movies_for_row } from './api/fetchMovies';
+import { Slider } from './components/Slider';
 
-document.querySelector<HTMLDivElement>('#app')!.innerHTML = `
-  <div>
-    <a href="https://vite.dev" target="_blank">
-      <img src="${viteLogo}" class="logo" alt="Vite logo" />
-    </a>
-    <a href="https://www.typescriptlang.org/" target="_blank">
-      <img src="${typescriptLogo}" class="logo vanilla" alt="TypeScript logo" />
-    </a>
-    <h1>Vite + TypeScript</h1>
-    <div class="card">
-      <button id="counter" type="button"></button>
-    </div>
-    <p class="read-the-docs">
-      Click on the Vite and TypeScript logos to learn more
-    </p>
-  </div>
-`
+const items_per_page = 6;
 
-setupCounter(document.querySelector<HTMLButtonElement>('#counter')!)
+// ✅ Everything initializes here
+window.addEventListener("DOMContentLoaded", () => {
+  document.querySelectorAll<HTMLElement>(".sliderWrapper").forEach(wrapper => {
+    const slider_content = wrapper.querySelector(".sliderContent") as HTMLElement;
+    fetch_movies_for_row(slider_content, items_per_page * 3).then(() => {
+      new Slider(wrapper, items_per_page);
+    });
+  });
+});
